Add consider-dev option for lock
This option attempts to resolve potential conflicts in the locked package versions emerging when dev packages require a lower version of a package than default packages.

Example problematic Pipfile:

[packages]
requests = "<2.20"
[dev-packages]
idna = "<2.7"

requests resolves to 2.19.1 which depends on idna <2.8,>=2.5. Normally, an idna version that satisfies both default and dev requirements would be for example 2.6.
However, since dev packages are not considered when resolving default packages, idna==2.7 gets selected as it is the highest version fulfilling default constraints.